修改了float.h中fraction的返回值类型,Integer -> Float。为了做类型统一。

+++ src/float.h
@@ -197,7 +197,7 @@ Float gamma(const Float& x, const char* epsilon=kDefEpsilon);
 //
 // 分数相关
 //
-std::pair<Integer, Integer> fraction(const Float& x);
+std::pair<Float, Float> fraction(const Float& x);
 
 //
 // 功能函数
